feat: store structured data in scrape report (#268)
Currently the content scraper returns only the body part of a website
and removes the head part including script tags, which can include
structured data.

To support the structured data audit use case, structured data
annotations in the head (in particular LD-JSON) needs to be extracted
and included in the crawling report.

* Adjusted default evaluation script to parse LD+JSON from script tags
and add it to the scrape reports.
* It will additional flatten `@graph` entries, so that parsed data has a
uniform structure.
* Minor fix in `local.js` to enable local execution again.

See SITES-28713 for more details.